Fixed StartBootstep2 handling again
Setting of ExpNmtState to PreOp2 is now only done if
current ExpNmtState is PreOp1. CNs which got an NMT ResetNode, because the configuration
has been restored are not in NodeStateUnknown, but in NodeStateRestoredConf, but they are
NotActive. Now there may be a race condition if NMT ResetNode has been issued, but ExpNmtState
is still PreOp1, because it hasn't been sent yet. How to solve this? Or is this race condition
not possible at all?