Adoptable Object Handling [8]: Use new kEplObdEvInitWriteLe event
kEplObdEvInitWriteLe is used in case of an little endian write access,
because then m_pData points to little endian data.
Currently, kEplObdEvInitWriteLe equals kEplObdEvInitWrite to be compatible
backwards. m_pData is only meaningful if the access is adopted and that is
only allowed if it is a little endian access. m_pArg points to the object size.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>