Adoptable Object Handling [9]: New error code kEplObdNumValSizeMismatch
Return kEplObdNumValSizeMismatch in case the segment size on a read access
does not match the object size if the object is a numerical object.
In previous releases it was allowed to pass a larger data buffer than the
object needs even for numerical objects. To have a minimal type safety the
data buffer size supplied by the caller has to match the object size in OD.
To distinguish this case from kEplObdValueLengthError a new error code is
introduced.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>