EplSyncu: Add guards to prevent compilation errors.
If EPL_DLL_PRES_CHAINING_MN is defined as FALSE in EplCfg.h, the stack won't compile.
The reason is, that parts of the stack (e.g. the type definition for tEplDllSyncRequest in EplDll.h) are guarded by EPL_DLL_PRES_CHAINING_MN, while the header and source file for EplSync (who use this type def) are not.