[FIX] QT demo misses CN reconnect
Sometimes when a CN was removed or reconfigured it would not shown
in QT demo when it came back. The CN was only added when
kEplNmtNodeEventFound was signalled but this is not always the case
(e.g. after the node was reconfigured). Now the node will also be added
when it reaches Preop1/Preop2/ReadyToOperate state.

Change-Id: Ie770b6eb143dfbc22b1756fe5682fa16c890c57c
Resolves: #31