[FIX] Obd/SdoSequ: Implement object 0x1300 (SDO Sequence layer timeout)
This commit adds an implementation for handling the
object dictionary entry 0x1300 (SDO Sequence layer timeout).
The default value used for this timeout is still 5000 ms as before,
but it can be changed now by writing object 0x1300.

The Ethernet POWERLINK specification states that this object shall
be 'valid on reset'. Thus, a new timeout value is only adopted
after a reset configuration event.

The timeout is internally limited to an upper bound of 86400000 ms,
which is exactly 1 day.

Change-Id: I68dd805edb04e431dcfcf1bdd6bd0f1cf7b0f6d6