[TASK] Adapt sources to new directory structure and naming conventions
This is the first big step for restructuring the openPOWERLINK tree.
Most of the source files are renamed and moved to the new directories.

Source code for architectures not maintained by B&R is moved into a
staging tree. It can be ported later or will be removed if it is no
longer maintained.

Makefiles will be adapted in further commits. Main header files will
also be restructured in further commits as this causes many changes
in the source files.

Change-Id: I67e3c921a932daa77d15552f64313a87cb804104
Reviewed-on: http://brateggevoat1/gerrit/553
Reviewed-by: Josef Baumgartner <josef.baumgartner@br-automation.com>
Tested-by: Josef Baumgartner <josef.baumgartner@br-automation.com>