[FEATURE] Add circular buffer library
This commit adds a circular buffer library. The library provides
a simpler mechanism than the existing circular buffers contained
in the sharedbuff library and should fully replace it in the
future. However, this library is not compatible with the existing
circular buffers. This means that the modules which should use
it, must be adapted to the new library. The library contains
implementations for different architecturs.

Change-Id: I26d099ad7ac05b3e94fa664a4a661771e8f08ba2