[TASK] Optimize asynchronous scheduling in DLLKCAL with circbuffer instances
In order to optimize the CN request handling of the asynchronous scheduler the
CN requests are forwarded to an NMT and GEN circbuffer instance. Note that
only the node ID is forwarded to the queue.
In order to get a pending CN request the queue is simply read, which reduces
the former worst-case iterations from 2*254 down to 2!
On Altera Nios II target the processing time is reduced from 75 us to 15 us!

Change-Id: I6e569a363a1a712a08072fab84dfed4f4d5301f6
Reviewed-on: http://brateggevoat1/gerrit/1314
Tested-by: Jenkins <jenkins@brateggevoat1.br-automation.co.at>
Reviewed-by: Josef Baumgartner <josef.baumgartner@br-automation.com>