[TASK] Update openCONFIGURATOR project Demo_3CN
Update the openCONFIGURATOR project for the demos. It is based on
the delivered CiA401_CN XDD file. The demos are adapted for the
changed xap.h file. The openCONFIGURATOR project was built with
openCONFIGURATOR 1.3.

The PResActPayloadLimit_U16 in the XDD files was adapted to a default
and maximum value of 40. This enhances compatibility for CNs which have
other default values for this object. Because the new default value
in the XDD differs from the actual value in the demo project it is
added to the CDC file by openCONFIGURATOR and therefore the CN is
correctly initialized.

Change-Id: Ia776d7579fd0d45fc7aad627577aa5d671cefd3f
Reviewed-on: http://brateggevoat1/gerrit/1383
Reviewed-by: Wolfgang Seiss <wolfgang.seiss@br-automation.com>
Reviewed-by: Joerg Zelenka <joerg.zelenka@br-automation.com>
Tested-by: Jenkins <jenkins@brateggevoat1.br-automation.co.at>
Reviewed-by: Josef Baumgartner <josef.baumgartner@br-automation.com>