[TASK] Split header files into internal and external files
Split header files into stack internal files and external files which
are needed by an application. Stack internal files are moved into
stack/include. The build system of the applications is adapted for the
new header file locations.

This is only the first step in cleaning up header files. More cleanup is
needed!

Change-Id: I625d5d22454c3614389637d8dfdc17da9f5152cb