[TASK] Remove dependancy of DLL configuration in dllcal-hostif.c
The code uses the configuration options CONFIG_INCLUDE_DLLU and
CONFIG_INCLUDE_DLLK to determine if it is compiled for the host (user)
part or the pcp (kernel) part. These are replaced by a separate
configuration macro CONFIG_HOSTIF_PCP which should be set to TRUE for
PCP and FALSE for Host.

Change-Id: If7420168db0d64c2a5e6c1b53fad389b415959be
Reviewed-on: http://brateggevoat1/gerrit/1751
Tested-by: Jenkins <jenkins@brateggevoat1.br-automation.co.at>
Reviewed-by: Joerg Zelenka <joerg.zelenka@br-automation.com>
Reviewed-by: Josef Baumgartner <josef.baumgartner@br-automation.com>