[TASK] Cleanup and restructure generic include files
The change cleans up the inclusion of files. The basic header file is
oplkinc.h. It ensures that all definitions are set, the generic and the
target specific ones. It includes all necessary files. It has to be
included by every other header file to ensure that the complete
environment is completely set.

- Rename Epl.h to oplk.h
- Rename EplInc.h oplkinc.h
- Move target specific definitions into separate header files in
  targetdefs
- Adapt all references

Change-Id: I13dae5ad497c5f77d500e28668d2b6ef63efb81e