[FIX] Stop configuration process on IdentResponse and NotActive
When the connection to a node is lost (because of NMT reset for
example), the MN indicates a NMT state change kEplNmtCsNotActive, but
a CFM configuration process for that node might still be running. The
same applies for reception of IdentResponse (kEplNmtNodeEventFound).
This commit closes a potentially running SDO connection on those
events.

Based on change for V1.08.4 of Daniel Krueger
<daniel.krueger@systec-electronic.com>

Change-Id: Ia86743dea134f2bc5c9b9f96bc61c8eef57b862c