[FEATURE] Add support to run demo_mn_embedded on target Zynq SoC.
The demo MN application for target Zynq SoC uses dual
processor stack library. A separate daemon application is
also needed to run this demo which should be compiled separately.

The application for Zynq SoC is modified to use configuration file
(mnobd.cdc) placed on the SD card along with binaries instead of
compiling it with source.