[FIX] Increase default value for request queues
On a network with more than 10 nodes the queue for status requests
was too small and buffer overflows occured. Therefore the size of
the request queues for IdentReq and StatusReq is increased to be
large enough for networks up to 40 nodes.

For small MN systems, the values can be scaled down by redefining
them in oplkcfg.h.

Change-Id: I2252db5175bec38274c095243beb557ac06267d7