[TASK] Add configuration of CFM in MN demos
The demo applications use the preprocessor macro CONFIG_INCLUDE_CFM
to determine if they shall contain CFM functionality. This macro
is defined in the oplkcfg.h header files which is no longer included
after clean-up of the include files. Therefore it is necessary to
configure it through the build system.

Change-Id: I4cae1194419354465a300b9cc7ae22597caef98a