[FIX] Fix include directories in Linux kernel driver
The CMake configuration for the Linux kernel driver is special
because it only links the source files into the build directory,
set compiler options and then uses kbuild for doing the real build.

Therefore, it is not necessary to set include directories with the
INCLUDE_DIRECTORIES statement. However, if a eclipse CDT project is
created by CMake the include directories must be correctly set for
the indexer to work.

This change correctly sets the include directories in the Linux
kernel module project.

Josef Baumgartner <josef.baumgartner@br-automation.com>
- Cleanup change and commit message for integration

Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>

Change-Id: I57bc721dc388971e2357b5675034914526488607