[FIX] Fix MN issuing StatusRequest after NMTEnableReadyToOperate
If the MN issues an NMTCommand NMTEnableReadyToOperate the initial and
destination NMT state of the node must be NMT_CS_PRE_OPERATIONAL_2.
Therefore, it is not necessary to issue a StatusRequest frame after
C_NMT_STATE_TOLERANCE cycles monitoring the state change.
This change skips setting up the state change monitoring if the
initial and destination state is NMT_CS_PRE_OPERATIONAL_2.

Resolves: #838

Change-Id: I09bb1bfe7998b323f36a47835926a2a2bcae4c94