[FEATURE] Add support for detecting kernel compatibility
This change adds support for detecting the kernel stack version and
configuration. The stack configuration is stored with a feature flag
variable that could be read by the user stack via the control interface.
The user stack compares the received feature flags and version of the
kernel stack with its expected flags and own kernel version to check
if they are compatible.

For checking the version and features, the command interface of the
control module is enhanced. Therefore, no new communication mechanism
is needed in the control CAL modules for exchanging the data.

Change-Id: I156193912bf5d631bfa950d3c0ee9bba43a32ce9