[FIX] Fix shutdown behaviour if kernel initialization fails
If edrv_init() fails and the startup is aborted, the shutdown of some
modules can cause crashes.

If there is an error on startup before dllucal_init() is called,
the function pointers will not be set. Therefore a call to
dllucal_exit() causes a crash. Now, the function pointers will
be checked before calling them.
dllkcal_exit() now also checks pointers before accessing it.

eventkcal-linuxkernel.c only waits for the eventThread to terminate
if it was really created.

Change-Id: Ic1ecf66e1117101b053924a031a5d1e24f13c5cd