[FIX] Fix resending of the same Tx buffer in BasicEthernet state
This spurious retransmission might lead to kEdrvNoFreeTxDesc with some
Ethernet drivers. A new state is needed which marks the Tx buffer as
being transmitted currently. So this buffer isn't retransmitted again
in dllk_processFillTx(). This new state is only important in NMT state
BasicEthernet or when entering this NMT state.

This change is based on change for V1.08.5 of
Daniel Krueger <daniel.krueger@systec-electronic.com>

Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>
Change-Id: I3593053442c7c69f94d08990e9d867ff238000ed