[TASK] Remove ftracedebug from demo applications
The ftracedebug module is only used on the Linux platform for easier
stack debugging. Therefore, it is included only in those stack files
referencing the FTRACE_* macros.

Additionally, all references to ftrace are removed from the demo
applications, and the ftrace initialization and cleanup are moved
from the application initialization to the stack target files.

Change-Id: I850f9f13406042ed66f20e951ca4f8d873eb1ea1