[FIX] Exit kernel event process in noos drivers
If the kernel event module is not initialized, calls to
eventkcal_process would try to access the event queues. In debug mode
every call creates an error message, leading to endless prinouts until
the user layer starts to initialize the kernel.
This fix avoids the endless printouts by returning from the process
function if the event module is not yet initialized.

Change-Id: I8230770858d5b7c467cd2925354c0ef6b5c545f9