[FEATURE] Add user CAL modules for Windows PCIe interface
The Windows PCIe user CAL uses a kernel driver as the communication
interface to the kernel stack running on a PCIe card.

The CAL interacts with the driver using IOCTLs. The IOCTL buffer is
used to pass additional information which is required for processing the
IOCTLs. The driver parses the IOCTLs and forwards the command, event or
data to the kernel stack through shared memory.

Change-Id: I769b2a09966c4df4380c61ee118553d966bbc405
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>