[FIX] Fix NULL pointer crash in edrvcyclic
At shutdown, functions of edrvcyclic could be called by other threads
even if edrvcyclic_shutdown() was called. This can cause a NULL
pointer access.

To avoid this race condition, and don't use any locks, the shutdown
of edrvcyclic module is moved after the shutdown of events and DLL.
In the past, edrvcyclic was shutdown by the DLL. To be able to change
the shutdown order, it is now directly called by the control module
together with the edrv module.

For a clean structure the initialization needs also be changed.
In the past, the MAC address was copied into the DLL module at
initialization. As the initialization of edrv and DLL is separated
now, the DLL needs to get the MAC address by calling an edrv
function insted of storing its own copy.

Change-Id: Ie09c658944aed2bad6421f7556ccd6a55a9ab9ac