[FIX] Fix calling of uninitialized mutex in pdou.c
If stack initialization fails before pdou_init() is called, the
mutex is not created. Therefore a call to pdou_exit() creates
an invalid memory access when trying to use the mutex. This change
creates an initialization flag which is used to detect if the
module needs to be cleaned up on pdou_exit().

Change-Id: Id4a3e7062e2ff78db1ab66fbae7505d8fd04a17d