[FIX] Fix crash with i210 if driver initialization fails
If driver initialization fails in initOnePciDev(), the function
calls immediately removeOnePciDev(). For the i210 the hres
timers are also using functions of the driver. Therefore they can
cause a crash if the driver isn't initialized. We protect
these functions by a simple flag to prevent access to the
uninitialized driver.

Change-Id: Ib7d766b2f8824c215bcb355b6628b86b037d0a3c