[FEATURE] DLL: Add RMN support
This includes the following features:
- support for Switch-over time
- AMNI receive frame handling
- transmission of AMNI on switch-over timeout (ASAP in timer callback
  function)
- signalling of active Managing Node Redundancy in SoA frame
- SoC relative time is kept running on switch-over

Some implementation comments:
1. The automatic frame transmission of edrvcyclic module has to be stopped, if
the node changes from Active Managing Node to Standby Managing Node (e.g.
upon SoC reception).

2. When entering Active Managing Node state, the SwitchOver timer isn't needed
anymore. It has to be freed so that the internal HighRes timer can be used
by edrvcyclic module.

3. Cycle preprocessing is triggered on switch from CS_ to MS_OPERATIONAL

4. Do not use LossOfSoC detection on SMN.
The Loss of SoC detection doesn't make sense on Standby Managing Node.
There the switch-over time shall be lower then object 0x1C14
DLL_CNLossOfSocTolerance_U32. Otherwise the Standby Managing Node might go
back to PreOp1, because of Loss of SoC threshold is exceeded. This lowers
the redundancy instead of directly entering AMN operational state.
Finally, this saves one high-res timer. Hence two high-res timers as most
of our target implementations can provide are still sufficient. When
running as SMN one is used for cyclic timer by edrvcyclic module and the
other is the switch-over timer.

5. Restart cyclic timer on each SoC reception.
This synchronizes the edrvcyclic module on the SMN to AMN, so it can switch
over to AMN while keeping the SoC timing. From network point of view, there
is an additional SoC delay between RMNs during switch-over which is induced
by this software solution. This is the receive-transmit round-trip time
(i.e. reception of SoC by Ethernet controller -> interrupt -> configuration
of high-res timer; occurrence of high-res timer -> interrupt -> SoC
transmission to Ethernet controller and then to wire). It might be
in incorporated by some kind of preshift in a future version.

6. Update CN MAC addresses from PRes if SMN.
The SMN might not receive the IdentResponse. So it needs another way to
update the MAC address of the other CNs for later use in PReq if it
switches over to AMN. Hence the MAC address is taken from the received
PRes.

7. Reset cycle counter properly for Loss of PReq detection.
Loss of PReq detection works only if cycle counter is zero. MN uses the
cycle counter for reduced cycle counting. So the cycle counter shall be
reset when AMN enters full cycle (PreOp2) and during switch from AMN to SMN
PreOp1.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>
Signed-off-by: Michael Ulbricht <michael.ulbricht@systec-electronic.com>