[FIX] Linux Mutex: Unlink semaphore before creating a new one
The semaphore might exist with the very same name from a previous
application run, but it might have the wrong init state (e.g. zero).
Then the first call to sem_wait will block infinitely.

Warning: This change assumes, that target_createMutex() is only called once
for each mutex. This is currently true, because the one and only user of
the mutex function class is user/pdo/pdou.c.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>