[FIX] Fix forwarding of wrong NMT states to the application
If a wrong NMT state is received in a StatusResponse or IdentResponse,
the stack detects that this is an invalid one, but it stores and
forwards it to the application. This can cause a crash if the
application does not check for a valid state.

Now, the NMT state received in StatusResponse and IdentResponse is
checked for valid states. If it contains garbage, a dummy state
(kNmtStateInvalid) is assigned. Additionally, the new state is only
forwarded and stored if it is a valid CN state.

Change-Id: I7d772a894d0f1feb9c4428b5ccbded66d492f549