[FIX] Discard invalid PDOs
If we get a PDO of a node which is not mapped, the lookup table is not
initialized and therefore points to the first PDO channel. On the
transmit side, those TPDOs are mapped to channel 0 by mistake and
filled with data not intended for it. To avoid using this invalid PDOs,
we compare the real node ID with the node ID of the PDO channel. 

Note: This is a workaround. For a clean fix the lookup table should
be changed so that we can detect if it is an unmapped PDO.

Change-Id: Icc5d539a427f34829c4ddb192e748ab24f50adcf
Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>
Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>