[FIX] Fix MN assigns unrequested async slots
Previously, the MN assigned slots to nodes that did not request for them
in the last received PRes frame. Usually this error happened in case of
low asynchronous load.
This bug is solved by decrementing the local request count after
updating it with the latest received request count (PRes frame).

Change-Id: I19dc29ff0abb5bdbef60c0ace47f5110c7adfa4e