[TASK] edrvcyclic_stopCycle() can optionally keep cycle timer running
This change is important for RMN.
The transmission of the current cycle is stopped always. This includes
the clearing of the entire Tx buffer list. So
edrvcyclic_setNextTxBufferList() works even if edrvcyclic_startCycle()
isn't called before.

Optionally, the cycle timer can be kept running. This ensures that the
cycle time is stable, even if the node switches from AMN to SMN and
back again (if other RMNs fail to become AMN).

Change-Id: I03018691968ad2ced233ea77af7febdb7ff19e3b
Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>
Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>