[TASK] DLL/RMN: Implement handling for NMT command GoToStandby
The switch-over timeout shall be started on last SoC transmission before
NMT command GoToStandby transmission. The implementation is a little bit
tricky, because the switch-over timer does not run on AMN. So we need to
start the timeout at the cycle after the last SoC transmission. Then the
actual timeout is one cycle length less.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>