[FIX] circbuf-linuxkernel: Use IRQ-aware spinlocks
On non-RT_PREEMPT kernels circbuf_lock() is called from IRQ context also.
There we need the IRQ-aware spinlocks. Otherwise the may spin infinitely in
the IRQ handler.

Note: circbuf_createInstance() creates a new spinlock for each instance,
even if it shall connect to an existing instance. Currently, this seems to
work, because the kernel part of the stack runs solely in the Linux kernel
space and doesn't use circbuf_connectBuffer().

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>