[FIX] Use IRQ-aware spinlocks in circbuf-linuxkernel.c
On non-RT_PREEMPT kernels circbuf_lock() is called from IRQ context
also. There we need the IRQ-aware spinlocks. Otherwise the may spin
infinitely in the IRQ handler.

Note: circbuf_createInstance() creates a new spinlock for each
instance, even if it shall connect to an existing instance. Currently,
this seems to work, because the kernel part of the stack runs solely
in the Linux kernel space and doesn't use circbuf_connectBuffer().

Change-Id: I648f478f03c218cdbcba41d06755e25465842421
Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>
Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>