[FEATURE] Multiple PDO channels
This change enables to use multiple PDO channels for a single frame.
This is necessary if more than 254 objects should be mapped. In this
case the mapping of a single frame is split into several PDO channels.

To be able to use multiple PDO channels, the PDO channel needs to be
enhanced to be able to store all necessary information. The copy
operations from/to the PDO triple buffer are modified and the PDO
lookup table is enhanced to be able to store several PDO channels for
a single node.

NOTE: The number of transmit PDO channels (D_PDO_TPDOChannels_U16) is
set to 1 in defaultcfg.h for CNs. If multiple transmit PDO channels
shall be used on a CN this number has to be increased.

Change-Id: Ib1bcc41aafb7ce171b4ccac88b63614b87b347dd