[TASK] Store sync response callback functions queue-based
The syncu module stores callback functions which are called when a sync
response frame is received. Previously, only one callback function
could be stored at once, which led to errors (kErrorNmtSyncReqRejected).

This commit introduces a queue-based storage of the registered callback
function. This allows to issue multiple sync requests to a single node.
The queue-based storage assures that the order of the requests is
preserved.

Change-Id: I282df51ad9bcfdeb0c26823f6cc11bfad00d18da