[FIX] Use a better regex for CMAKE_SYSTEM_PROCESSOR
According to CMake doc, CMAKE_SYSTEM_PROCESSOR is set to the
output of the command "uname -p" (in fact it is the output of "uname
-m"). This is only true for native compilation.

In the case of cross-compilation, the CMake code does not set anything,
it just expects the CMAKE_SYSTEM_* variables to be set in the
toolchain file.

Add a test matching the patern "^i.86$" and keep the old one "x86" for
compatibility.
By doing that, factorize this test using "^(i.86|x86(_64)?)$".

Thanks to Samuel Martin
http://lists.busybox.net/pipermail/buildroot/2014-November/112163.html

Change-Id: If8823331280ed9b01b285bb27a0a3f071ab1c7e5
Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Josef Baumgartner <josef.baumgartner@br-automation.com>