[FIX] Fix for sync irq miss on c5SoC ARM during critical sections
Only disable the irqs on the ARM cpu interface when interrupts
are disabled/enabled during critical sections. This ensures that
the cpu is not interrupted during critical sections.

Enable the irq on the global interrupt distributor interface only at the
initialization and disable it during deinitialization. This ensures that
the irqs generated during critical sections are serviced by the cpu
after the irq on the cpu interfaced is re-enabled.

This fixes the sync irq drops on the host c5SoC ARM processor.

Change-Id: I2759c1c72f7291220abbeb7f4c53f09726a54fce