[TASK] Rework kernel event handling for FPGA drivers
This commit reworks the event handling in the kernel layer. The
objective of this rework is to avoid shared locking which blocks the
kernel layer from posting events into the kernel-to-user queue, leading
to cycle time violation.
In order to work around shared blocking the events intended for the
user layer are posted to the kernel internal event queue instead. The
background task then posts events intended for the user layer to the
kernel-to-user queue without disabling the global interrupt. This
enables any interrupt to suspend the queue handling.

Change-Id: Ifb9423199fec0b56dcddb745f04f5057645d0cbd