[FIX] Fix asynchronous Rx count in DLLKCAL
The asynchronous Rx count was modified in interrupt context and also
from the background loop. This could lead to corruption of the count
value fooling the asynchronous scheduler, which then stops to allocate
the asynchronous slots.
This fix introduces two independent counter variables which track the
received and freed asynchronous Rx frames. In order to get the number of
asynchronous Rx frames in the queue the difference is calculated.

Change-Id: Id01f334d43cea52ffdc66daac4397f64dc96bedf