[FIX] Reduce timeout steps in ctrlucal_executeCmd
In the case of an hostinterface CAL, the function
ctrlucal_executeCmd() posts a command, then checks periodically for a
result before possibly exiting because of a timeout.

This patch reduces the granularity of those checks from 1 second to
10 milliseconds. Since the functions starts by waiting one period,
and is called several times during the initialization, this patch
saves around 5 seconds of boot time in the case of dual-core host
interface designs.

Change-Id: I16548676260f992dad29085ed822182da47a8880
Signed-off-by: Sylvain Lesne <lesne@alse-fr.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>