[TASK] Rework openMAC second timer resources
The second timer in openMAC previously generated a toggling signal in
case of timer value match.
Now, the second openMAC timer is reworked to generate a pulse for host
processor synchronization. This pulse lasts for one clock cycle (20 ns)
which is compatible to the host interface ipcore synchronization input.
Optionally the pulse width can be controlled if the pulse width control
register is enabled in the ipcore settings - this is useful if the pulse
needs to be detected with a lower sample rate than the pulse duration.

Change-Id: I215907407ca7398a565f2bc9585a22fee969c1b0