[FIX] demo_mn_qt: fix regression in the SDO dialog
Commit 86c8846 created a nice SDO dialog in the Qt MN demo, but
commit 275bd94 (which introduced fixed-size data types to avoid
using enumerator types across the CAL) initiated a regression: any
attempt at doing an SDO access will lead to "Error 0x142: Invalid
parameters were passed to a function (e.g. invalid node id)".

This is related to a cast within the QVariant class. In this case,
the QVariant holds an enum (since it was created as such:
`QVariant::fromValue(kObdTypeUInt8)`), and doesn't seem to want to
convert this to an integer, returning the default int value (0).
This causes the infamous error 0x142.

So, this patch makes sure that enumerator types are used all the way
with QVariant instances, and casts them back to the related
fixed-size data type when calling an OPLK library function.

Change-Id: I601f36b6cee37eb12d8fad63f1c6c1bc00c3afa4
Signed-off-by: Sylvain Lesne <lesne@alse-fr.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>