[FIX] Fix DE2-115 SRAM timing issues
This fix solves the Nios II ELF verify issues by improving the I/O
timing of the tri-state-bridge connected to the SRAM. This is achieved
by forcing the fitter to use fast I/O registers available in the FPGA's
IOBs. Additionally the maximum tco is reduced raising timing violations
if the fast I/O registers are not assigned by the fitter. Also enable
maximum current strength of control signals for best interface
performance.

Change-Id: I0d8f00031fcc8c4f0f0d8c810e2f26fd316bcdb0