[FEATURE] Add dualprocshm support files for PCIe device with NIOS2
This adds support for systems consisting of an x86 processor running
Windows as host processor and Nios II soft-core processor on a PCIe
card as driver.

The memory shared between the processors is accessed over PCIe.
The memory on the PCIe card is mapped for Windows through 2 BARs. The
dualprocshm uses BAR0 as the shared memory and BAR1 as the common
memory.

This commit adds:

- Target support header for Windows.
- Target support modules for PCP daemon and PCIe interface.

Change-Id: I2abdee0c0b9ee7ddfcda110a0dd7c45325294481
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>