[FIX] Fix CFM feature mismatch between demo and stack
The demo applications contain some code used for initialization of
objects if the stack is compiled without the CFM feature. However,
the demos are targeted for a CFM enabled stack and do not work
without CFM. If the application is compiled without CFM function but
the stack includes the CFM feature this results with a wrong
initialization of nodes. Therefore, all application code which is
targeted for a non CFM stack is removed. The configuration option
for the CFM in the demos is also removed.

Change-Id: I38d5439966d57ca7ccf9ea49b839a70566bcd094