[TASK] Acknowledge openMAC synctimer at beginning of ISR
Previously, the openMAC synctimer IRQ was only acknowledged in
drvConfigureShortestTimer() which is called at the end of the ISR.
If the IRQ deassertion delay is too high, the ISR is re-entered by
mistake.
By adding an additional IRQ acknowledgement at the beginning of the ISR,
it is less likely that this error will occur.

Change-Id: I0e9d1c4cf24ea99e8b4ea4b0a2533ee89a750e9b