[FEATURE] Add eventlog library
The eventlog library provides functionality for openPOWERLINK
applications to log openPOWERLINK events in a unified format.
It contains two modules:
- eventlogstring.c: Low level module to create the unified log message
                    strings.
- eventlog.c:       High-level module to be used by the application for
                    printing the log files.
It supports printing in two formats. One format is intended to be
parsable by tools for evaluation and testing. The second one is
formated in a more readable way for humans. A filter can be
specified to allow filtering on specific log levels or log categories.

Change-Id: Ie1e9e0d316bbc70a4180d89bd1bac3b1fe172bd6