[FEATURE] Rework demo_cn_console to use new event log library
demo_cn_console is changed to use the new event log library. Command
line options for the output format and the log filter are added.
Some basic information output is done via printf as the log output goes
to standard error and can be redirected and filtered.

The unsused option to specify a log file is removed.

Change-Id: Ia36c1a098777863b93655722bf4e1345d639c841