[FIX] Fix loss of events in Windows PCIe solution
The event CAL module for Windows PCIe solution used
two threads to handle events received from the kernel-to-user and
the user-internal queue. One thread was used as a common thread
for processing the received events and the other thread to retrieve the
events from the kernel-to-user queue.

In case of an increased system load it was noticed that events received
from the kernel-to-user queue were dropped due to synchronization errors
between the two threads. The system load delayed the execution of the
common thread causing e.g. configuration download failures and NMT state
change errors.

This commit revises the event handling to use a single thread for event
processing and retrieval. The thread polls the user-internal and
kernel-to-user queues for new events and processes them sequentially.
Since polling and processing is done in a sequential order, even though
the increased system load will delay event processing, it won't result
in event losses.

Change-Id: Ic90d7d6e410b419650f913199763456556552051
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>