[FIX] Fix loss of events in Windows PCIe solution
The event CAL module for Windows PCIe solution used
two threads to handle user and kernel events processing.
The first thread is used as common thread to process the
events and second thread to retrieve kernel to user events
from kernel layer.

In this two thread implementation, with increase in the system
load, it was noticed that kernel to user events were dropped
due to synchronization errors between the threads.

Increased system load delayed the execution of the common
thread due other higher priority system threads and
resulted in loss of kernel events. This caused issues
such as configuration download failure and NMT state change
failure.

This commit revises the event handling to use only single
thread for event processing and retrieval.

The thread polls kernel to user and user internal
queues for new events and processes them sequentially.
As the polling and processing in done in a sequential order,
increased system load delays the processing of the events
but does not result in event loss.

Change-Id: Ic90d7d6e410b419650f913199763456556552051
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>