[FIX] Fix header inclusion order to avoid compilation warning in obd.h
common/oplkinc.h is included by default in all headers. It also
includes oplk headers including obd.h and oplkcfg.h. This produces
compilation warnings when any stack configuration change in oplkcfg.h
is not reflected in obd.h because of the order of inclusion. So, the
order was changed to configure the oplk headers as per the
configuration in oplkcfg.h.

The dualprocshm-mem headers are included in oplkcfg.h and use the
exported defines from the hardware headers. To avoid compilation
error due to the change of inclusion order, the hardware exported
headers are included in the dualprocshm-mem headers.

Change-Id: Id2fd86973eccd4b699edc382a4e248fd3309026e
Signed-off-by: John <chidrupaya.sr@kalycito.com>