[FIX] Update dualprocshm driver library configuration settings
The PCIe design does not support VEth interface currently.
Due to this the openPOWERLINK kernel library for dualprocshm
interface should be compiled without the VEth.

As a workaround the VEth module was removed from the kernel
configuration, which caused feature mismatch on Zynq and
Altera SoC design.

The commit adds conditional compilation using a platform specific
compiler flag to identify PCIe design and disable the VEth
only for PCIe designs.

Change-Id: I55d096032bbb4fd5593412d83936ef48dd73d2d2
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>