[FEATURE] Improve display of LEDs in demo_mn_qt
Currently the LED images of the demo are displayed, even if the inputs
are not valid or the output data is not transferred into the network.
This change improves the demo so that the LEDs are only painted
if the node is an operational state and the output data is sent into
the network. The input LEDs are only painted if the input data is
processed. To show unused input and output data, the LEDs are painted
in gray color.

Additionally, false/true is used for setting bool
variables instead of the openPOWERLINK TRUE/FALSE macros.

Change-Id: I054f4a49204b8ebbad2675dfea5df0727764ab0e