[FIX] Fix for 911ed4175f8968dbf106e7b7e0853f28b547e346
New commit message:
[FIX] Avoid failure of embedded demos on platforms with no VEth

The embedded MN/CN demo fails on platforms which do not support
virtual Ethernet interface.

As virtual Ethernet is not a mandatory requirement, this fix
updates the embedded demo to avoid stack shutdown on such
platforms and displays a warning message instead on failure
to register the virtual Ethernet callback.

Change-Id: I0e3108f1145315969f29e3b1ed2c35dd71cf866d
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>