[FIX] Avoid error in embedded demos if virtual Ethernet is disabled
The embedded MN/CN demo fails on platforms which do not support
virtual Ethernet interface.

As virtual Ethernet is not a mandatory requirement, this fix
updates the embedded demo to avoid stack shutdown on such
platforms and displays a warning message instead on failure
to register the virtual Ethernet callback.

Change-Id: I59f1d587b92189155073b8f787cc8572a43d9263
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>