[TASK] Update openMAC edrvcyclic to application period configuration
The application can configure the minimum synchronization period it
supports. Introduced with this commit the configuration is forwarded
to the openMAC specific edrvcyclic module. This allows to set up the
external synchronization interrupt depending on this configuration.
The general edrvcyclic module simply ignores that parameter since it
is not responsible for synchronizing the application.

Change-Id: Iaadc282b140c06f8738d6039eadddea83bd4b6d6