[FIX] Fix for 574095ad587235c9897525dfa3496d3bb4e4eea3
New commit message:
[TASK] Add default PDO setup wait time constant

The PDOU module waits for the PDOK module to complete the PDO
buffer setup initialization in kernel space before initializing
the PDO memory in user space. This is done using a target
sleep routine with a constant sleep time on all platforms.

This commit adds a configuration constant with the default value
of 500 ms for the sleep.

Platforms where the wait time is different, this value can be
overridden by configuring the wait time constant in the
oplkcfg.h for the design.

Change-Id: I20bd6cb338fb35332a2fb4a931e7f4896011298d
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>