[FIX] Update call sequence of module exit during shutdown
On platforms which use a separate thread for handling events,
such as Linux and Windows, the event thread can process the
events even after the the dependant modules have already exited.
Any calls to the dependant modules for handling events after
the module exit results in application crash.

This commit updates the call sequence during the shutdown of
openPOWERLINK stack, so that the eventu module is exited
before dllu and timeru modules avoiding application crash
identified on Windows PCIe design with APC2100.

Change-Id: I0866453bc61d8ae1bbe5f3ecac67da4bd12d6179
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>