[FIX] Add initialization flag for eventu module
The resources for event handling in the CAL are freed during
the call to eventu_exit(). Any posts or processing requests
after the exit will lead to application crash.

This commit adds an initialization flag to determine the status
of the module and avoids processing of event post and process
requests.

Change-Id: I932f1e3ed78e77b35792893bbd02adf17cf52e20
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>