[FIX] Fix stack initialization error with Linux PCIe
The virtual Ethernet feature is disabled by default in Linux PCIe stack
library in order to match with the default feature setup of the kernel
stack on PCIe card. At initialization an error occurs since the kernel
space module has the virtual Ethernet feature enabled.
With this commit the feature is disabled also in the kernel space
module.

Change-Id: Iab9eebbaeec0ae7d5f211e2d93f3df582c34438c
Signed-off-by: Varun Rajendran <varun.ar@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>