[FIX] Adapt sequence layer timeout to EPSG DSP 301 V1.2.0
Instead of the old specification V1.1.0, the sequence layer
timeout value is increased from 5ms to 15ms, but the
sub-timeouts where an ack-request is sent uses a constant divider
(3). Therefore 2 ack-requests will be sent before the final
close, resulting in a final timeout of 15ms.
The final timeout can be set dynamically with object 0x1300/0x00.
The ack-request retry counter object 0x1302 is not implemented
for now.

Change-Id: I9ee5d41129bc646f81047db0aa0c9b72db9201ac