[FEATURE] Enable exception stack for Nios II MN driver
In order to accelerate interrupt routine processing the separated
exception stack supported by Nios II is enabled. The exception stack is
linked to the tightly-coupled memory region, hence it is outside of the
cachable region. Note that the driver must be compiled in release mode
to enable the exception stack feature.

Change-Id: I6eef822128a257e0f53498eb597f124381d21b81