[TASK] Mitigate feature check at stack start
Previously, the feature check which is done at stack start checks for
perfect match of the kernel and user feature flags. This check is too
restrictive since it requires to recompile the kernel stack
unnecessarily if a kernel stack feature is not used by the user stack.

Now, the feature check only fails if needed kernel stack features are
missing. If kernel stack features are available but not supported by the
user, the check does not fail.

Change-Id: I917995e15af46a587333ea49dda6ea5701e6a5ac