[FIX] Fix full kernel-internal queue on FPGA MN targets
The kernel-internal event queue is used on FPGA MN targets to shift
uncritical event processing (such as filling asynchronous Tx buffers)
to the background. For instance loss of PRes error events are
posted to the kernel-internal event queue which potentially fill up
the queue.
In order to handle large network configurations the kernel-internal
event queue size is increased - now configured to handle the maximum
possible loss of PRes count.

Change-Id: I9958715d1b3f234e3bd41f8f74f5026e9e6c3d16