[FIX] Fix startup problem with CNs using PRes Chaining
The MN forwards the PResTimeFirst value to the CN with the SyncReq
frame. The MN sends another SyncReq frame to the CN to verify if the
time has been taken by the CN successfully. Otherwise the MN issues a
reset node command and restarts the procedure again.
Since the CN only updates the SyncRes Tx buffer if the PResTimeFirst
changes compared to the locally stored the CN is not responding the
valid PResTimeFirst value.
This fix commit ensures that the SyncRes Tx buffer is updated with
the received PResTimeFirst even though the value has not changed.

Change-Id: Ie86089e63e9652c0fefac0cfdda4a96959365a9d