[TASK] Remove openMAC "no filter match" IRQ
The openMAC IP-Core does not support the no filter match IRQ and
openPOWERLINK is not using it. Therefore the functionality is removed
from the omethlib driver.
Additionally the Rx lost flag in the Rx register is moved to a
different bit position, which allows to increase the descriptor
pointer width without changing the register map.

Change-Id: I8a76ec3f54fe8f54a819310f4ab85f1bd7d5ed31