[TASK] Enable coexistence of openMAC TTTX and auto response
Previously, enabling the time-triggered Tx and auto response features
altogether reduced the number of available Tx descriptors for an MN.
This reduces the number of CNs which can be used in the network.
With the recently added Tx descriptors it is now possible to allow
the auto response feature along with the time-triggered Tx
capability.

Change-Id: I0a76c95788a5390cb3852c794861e3288285aa35