[FIX] Fix Zynq Linux MN wrong cycle time issue for large cycle times
The variable for holding the converted timer ticks from the
corresponding nano seconds of time was trimming the tick count to
16 bit. So when this trimmed 16 bit value was checked against the
maximum allowed 16 bit tick count by TTC, the error condition for
exceeding timer upper limit was never generated. This was causing an
automatic rollover of the timer values greater than the maximum
possible 16 bit timer value without throwing an error.

Extend the variable to hold the largest probable timer tick value from
the nano seconds passed without trimming it down. So when compared
against the upper limit for the timer tick, it throws an error without
causing an automatic rollover.

Change-Id: I5a9186a6730dbaf81be4e5c5fd187fc7789ea13c
Signed-off-by: John <chidrupaya.sr@kalycito.com>
Signed-off-by: Joerg Zelenka <joerg.zelenka@br-automation.com>